+++ app/src/main/java/eu/faircode/email/DaoFolder.java
@@ -120,8 +120,11 @@ public interface DaoFolder {
             " GROUP BY folder.id")
     LiveData<List<TupleFolderEx>> liveUnified(String type);
 
-    @Query("SELECT account, id AS folder, unified, sync_state FROM folder" +
-            " WHERE sync_state IS NOT NULL" +
+    @Query("SELECT folder.account, folder.id AS folder, unified, sync_state" +
+            " FROM folder" +
+            " JOIN account ON account.id = folder.account" +
+            " WHERE account.`synchronize`" +
+            " AND sync_state IS NOT NULL" +
             " AND folder.type <> '" + EntityFolder.OUTBOX + "'")
     LiveData<List<TupleFolderSync>> liveSynchronizing();
 
@@ -178,6 +181,7 @@ public interface DaoFolder {
     List<EntityFolder> getChildFolders(long parent);
 
     @Query("SELECT folder.type" +
+            ", COUNT(DISTINCT folder.id) AS folders" +
             ", COUNT(message.id) AS messages" +
             ", SUM(CASE WHEN NOT message.ui_seen THEN 1 ELSE 0 END) AS unseen" +
             ", CASE WHEN folder.account IS NULL THEN folder.sync_state ELSE NULL END AS sync_state" +