+++ app/src/main/java/eu/faircode/email/FragmentPro.java
@@ -59,11 +59,13 @@ public class FragmentPro extends FragmentBase implements SharedPreferences.OnSha
     private TextView tvList;
     private Button btnPurchase;
     private ImageView ivExternal;
-    private TextView tvNoPlay;
     private TextView tvPrice;
+    private TextView tvGoogle;
+    private TextView tvNoPlay;
     private TextView tvPriceHint;
     private TextView tvFamilyHint;
     private TextView tvRestoreHint;
+    private Button btnSupport;
     private Button btnConsume;
     private ImageView ivConnected;
 
@@ -87,11 +89,13 @@ public class FragmentPro extends FragmentBase implements SharedPreferences.OnSha
         tvList = view.findViewById(R.id.tvList);
         btnPurchase = view.findViewById(R.id.btnPurchase);
         ivExternal = view.findViewById(R.id.ivExternal);
-        tvNoPlay = view.findViewById(R.id.tvNoPlay);
         tvPrice = view.findViewById(R.id.tvPrice);
+        tvGoogle = view.findViewById(R.id.tvGoogle);
+        tvNoPlay = view.findViewById(R.id.tvNoPlay);
         tvPriceHint = view.findViewById(R.id.tvPriceHint);
         tvFamilyHint = view.findViewById(R.id.tvFamilyHint);
         tvRestoreHint = view.findViewById(R.id.tvRestoreHint);
+        btnSupport = view.findViewById(R.id.btnSupport);
         btnConsume = view.findViewById(R.id.btnConsume);
         ivConnected = view.findViewById(R.id.ivConnected);
 
@@ -158,6 +162,13 @@ public class FragmentPro extends FragmentBase implements SharedPreferences.OnSha
             }
         });
 
+        btnSupport.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Helper.view(v.getContext(), Helper.getSupportUri(v.getContext()), false);
+            }
+        });
+
         btnConsume.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -180,11 +191,13 @@ public class FragmentPro extends FragmentBase implements SharedPreferences.OnSha
         btnPurchase.setEnabled(!play);
         ivExternal.setVisibility(play ? View.GONE : View.VISIBLE);
         tvPrice.setVisibility(View.GONE);
+        tvGoogle.setVisibility(play ? View.VISIBLE : View.GONE);
         tvNoPlay.setVisibility(
                 BuildConfig.PLAY_STORE_RELEASE && !Helper.hasPlayStore(getContext())
                         ? View.VISIBLE : View.GONE);
         tvFamilyHint.setVisibility(play ? View.VISIBLE : View.GONE);
         tvRestoreHint.setVisibility(play ? View.VISIBLE : View.GONE);
+        btnSupport.setVisibility(play ? View.VISIBLE : View.GONE);
         btnConsume.setEnabled(false);
         btnConsume.setVisibility(ActivityBilling.isTesting(getContext()) ? View.VISIBLE : View.GONE);
         ivConnected.setVisibility(View.GONE);